+++ Cargo.toml
@@ -67,6 +67,7 @@ serde = { version = "^1.0", features = ["derive"], optional = true }
 sha2 = "^0.9"
 structopt = "0.3"
 tectonic_bridge_core = { path = "crates/bridge_core", version = "0.0.0-dev.0" }
+tectonic_docmodel = { path = "crates/docmodel", version = "0.0.0-dev.0" }
 tectonic_engine_bibtex = { path = "crates/engine_bibtex", version = "0.0.0-dev.0" }
 tectonic_engine_xdvipdfmx = { path = "crates/engine_xdvipdfmx", version = "0.0.0-dev.0" }
 tectonic_engine_xetex = { path = "crates/engine_xetex", version = "0.0.0-dev.0" }
@@ -127,6 +128,7 @@ tectonic_bridge_harfbuzz = "2c1ffcd702a662c003bd3d7d0ca4d169784cb6ad"
 tectonic_bridge_icu = "2c1ffcd702a662c003bd3d7d0ca4d169784cb6ad"
 tectonic_cfg_support = "thiscommit:aeRoo7oa"
 tectonic_dep_support = "5faf4205bdd3d31101b749fc32857dd746f9e5bc"
+tectonic_docmodel = "cd77b60d48b1ae3ef80d708e6858ea91cd9fa812"
 tectonic_engine_bibtex = "thiscommit:2021-01-17:KuhaeG1e"
 tectonic_engine_xdvipdfmx = "7dcbc52e58f9774b3d592919a9105377faeac509"
 tectonic_engine_xetex = "b7a4085fa67c831d4532da6661bddafd1f9c24ff"
