+++ src/bin/tectonic/compile.rs
@@ -18,9 +18,9 @@ use tectonic::{
     config::PersistentConfig,
     driver::{OutputFormat, PassSetting, ProcessingSessionBuilder},
     errmsg,
-    errors::{ErrorKind, Result},
+    errors::Result,
     status::StatusBackend,
-    tt_error, tt_note,
+    tt_note,
     unstable_opts::{UnstableArg, UnstableOptions},
 };
 
@@ -198,30 +198,6 @@ impl CompileOptions {
         };
         sess_builder.build_date(build_date);
 
-        let mut sess = sess_builder.create(status)?;
-        let result = sess.run(status);
-
-        if let Err(e) = &result {
-            if let ErrorKind::EngineError(engine) = e.kind() {
-                let output = sess.get_stdout_content();
-
-                if output.is_empty() {
-                    tt_error!(
-                        status,
-                        "something bad happened inside {}, but no output was logged",
-                        engine
-                    );
-                } else {
-                    tt_error!(
-                        status,
-                        "something bad happened inside {}; its output follows:\n",
-                        engine
-                    );
-                    status.dump_error_logs(&output);
-                }
-            }
-        }
-
-        result.map(|_| 0)
+        crate::v2cli::run_and_report(sess_builder, status)
     }
 }
