+++ src/bin/tectonic/v2cli.rs
@@ -10,11 +10,13 @@ use tectonic::{
     self,
     config::PersistentConfig,
     ctry,
-    errors::{Result, SyncError},
+    docmodel::{DocumentExt, DocumentSetupOptions, WorkspaceCreatorExt},
+    driver::ProcessingSessionBuilder,
+    errors::{ErrorKind, Result, SyncError},
     status::{termcolor::TermcolorStatusBackend, ChatterLevel, StatusBackend},
     tt_error, tt_note,
-    workspace::{self, Workspace},
 };
+use tectonic_docmodel::workspace::{Workspace, WorkspaceCreator};
 use tectonic_errors::Error as NewError;
 use tectonic_status_base::plain::PlainStatusBackend;
 
@@ -163,15 +165,33 @@ impl BuildCommand {
         let ws = Workspace::open_from_environment()?;
         let doc = ws.first_document();
 
+        // XXX NO WAY TO DISABLE INSECURE FEATURES
+        let mut setup_options = DocumentSetupOptions::new(false);
+        setup_options.only_cached(self.only_cached);
+
         for output_name in doc.output_names() {
-            let mut opts = doc.build_options_for(output_name);
-            opts.format_cache_path(config.format_cache_path()?)
-                .only_cached(self.only_cached)
+            let mut builder = doc.setup_session(output_name, &setup_options, status)?;
+
+            builder
+                .format_cache_path(config.format_cache_path()?)
                 .keep_intermediates(self.keep_intermediates)
                 .keep_logs(self.keep_logs)
-                .print_stdout(self.print_stdout)
-                .open(self.open);
-            doc.build(output_name, &opts, status)?;
+                .print_stdout(self.print_stdout);
+
+            run_and_report(builder, status)?;
+
+            if self.open {
+                let out_file = doc.output_main_file(output_name);
+                tt_note!(status, "opening `{}`", out_file.display());
+                if let Err(e) = open::that(&out_file) {
+                    tt_error!(
+                        status,
+                        "failed to open `{}` with system handler",
+                        out_file.display();
+                        e.into()
+                    )
+                }
+            }
         }
 
         Ok(0)
@@ -264,11 +284,42 @@ impl NewCommand {
             self.path.display()
         );
 
-        let wc = workspace::WorkspaceCreator::new(self.path);
+        let wc = WorkspaceCreator::new(self.path);
         ctry!(
-            wc.create(&config, status);
+            wc.create_defaulted(&config, status);
             "failed to create the new Tectonic workspace"
         );
         Ok(0)
     }
+}
+
+pub(crate) fn run_and_report(
+    sess_builder: ProcessingSessionBuilder,
+    status: &mut dyn StatusBackend,
+) -> Result<i32> {
+    let mut sess = sess_builder.create(status)?;
+    let result = sess.run(status);
+
+    if let Err(e) = &result {
+        if let ErrorKind::EngineError(engine) = e.kind() {
+            let output = sess.get_stdout_content();
+
+            if output.is_empty() {
+                tt_error!(
+                    status,
+                    "something bad happened inside {}, but no output was logged",
+                    engine
+                );
+            } else {
+                tt_error!(
+                    status,
+                    "something bad happened inside {}; its output follows:\n",
+                    engine
+                );
+                status.dump_error_logs(&output);
+            }
+        }
+    }
+
+    result.map(|_| 0)
 }
