+++ src/lib.rs
@@ -78,16 +78,13 @@
 mod app_dirs;
 pub mod config;
 pub mod digest;
-#[cfg(feature = "serialization")]
-pub mod document;
+pub mod docmodel;
 pub mod driver;
 pub mod engines;
 pub mod errors;
 pub mod io;
 pub mod status;
 pub mod unstable_opts;
-#[cfg(feature = "serialization")]
-pub mod workspace;
 
 // Note: this module is intentionally *not* gated by #[cfg(test)] -- see its
 // docstring for details.
